Adding Tests for Square

+++ Polygons.Library/Square.cs
@@ -8,7 +8,7 @@ namespace Polygons.Library
 {
     public class Square : ConcreteRegularPolygon
     {
-        public Square(int length) : base(4, length) { }
+        public Square(double length) : base(4, length) { }
 
         public override double GetArea()
         {
